# librefit-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: librefit-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
...
  /api/user/activate/{activationId}:
    get:
      tags:
      - User Resource
      operationId: activateUser
      parameters:
      - name: activationId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  /api/user/login:
    post:
      tags:
      - User Resource
      operationId: postUserLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibreUser'
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  /api/user/logout:
    post:
      tags:
      - User Resource
      operationId: postUserLogout
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
        "401":
          description: Not Authorized
      security:
      - SecurityScheme: []
  /api/user/read:
    get:
      tags:
      - User Resource
      operationId: readUserInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibreUser'
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/user/refresh:
    post:
      tags:
      - User Resource
      operationId: postUserRefresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInfo'
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /api/user/register:
    post:
      tags:
      - User Resource
      operationId: postUserRegister
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibreUser'
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
  /api/user/update:
    post:
      tags:
      - User Resource
      operationId: updateUserInfo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibreUser'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibreUser'
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
components:
  schemas:
    AuthInfo:
      required:
      - accessToken
      - refreshToken
      type: object
      properties:
        accessToken:
          type: string
        accessExpires:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
        refreshToken:
          type: string
        refreshExpires:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
...
    ErrorDescription:
      required:
      - field
      - message
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      required:
      - errors
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDescription'
...
    LibreUser:
      required:
      - email
      - password
      - role
      type: object
      properties:
        id:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        email:
          type: string
        password:
          minLength: 1
          type: string
        role:
          type: string
        name:
          type: string
          nullable: true
        registered:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
        lastLogin:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
        avatar:
          type: string
          nullable: true
        activated:
          type: boolean
...
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT